- Added common and contract-api modules, so that modules can use common utils, classes and also interact using interfaces in contract-api. - Added app module for main() start after building and compiling all the existing modules (dependencies in pom). Also it contains application.yaml centralized config, root db changelog master, that refers other modules' own changelog master files. - Updated identity pom and media pom, adding these modules as dependencies. - Updated root pom, adding all the containing modules there (and also adding empty reactions, comments and views modules. - Updated and moved to common module StringHelperUtils.
