+++ app/src/main/java/ua/nin/AppApplication.java
@@ -1,13 +1,13 @@
-package ua.nin.identity;
+package ua.nin;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class IdentityApplication {
+public class AppApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(IdentityApplication.class, args);
+        SpringApplication.run(AppApplication.class, args);
     }
 
 }
