+++ app/src/main/resources/db/changelog/db.changelog-master.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<databaseChangeLog
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog
+        http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.25.xsd">
+
+    <include file="classpath:db/changelog/db.identity-changelog-master.xml"/>
+    <include file="classpath:db/changelog/db.media-changelog-master.xml"/>
+    <include file="classpath:db/changelog/db.reactions-changelog-master.xml"/>
+    <include file="classpath:db/changelog/db.comments-changelog-master.xml"/>
+    <include file="classpath:db/changelog/db.views-changelog-master.xml"/>
+
+</databaseChangeLog>
+
